move to components directory and fix issue when name is undefined
